check for run file before deleting

+++ greencandle/bin/get_data.py
@@ -89,7 +89,8 @@ def main():
     redis = Redis(interval=interval, test=False)
     redis.clear_all()
     del redis
-    os.remove('/var/run/gc-data')
+    if os.path.exists('/var/run/gc-data'):
+        os.remove('/var/run/gc-data')
     prod_initial(interval) # initial run, before scheduling begins
     Path('/var/run/gc-data').touch()
     LOGGER.info("Finished initial prod run")
